Version09/pyproject.toml
[project]
name = "version09"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "akasha-terminal",
    "ipykernel>=6.29.5",
]

[tool.uv.sources]
akasha-terminal = { git = "https://github.com/iii-org/akasha.git", rev = "ceres" }
